add PinsListWidget
